backend/pyproject.toml
[project]
name = "kbalyzer"
version = "0.1.0"
description = "Kombuchalyzer backend, an open source API for Kombucha brewing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.31.0",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.3.0",
    "python-multipart>=0.0.20",
    "bcrypt==4.0.1",
    "qrcode>=8.2",
    "pyotp>=2.9.0",
    "pillow>=12.0.0",
]

[tool.ruff]
line-length = 120
exclude = [
    "./alembic",
]

[tool.ruff.lint]
select = ["ALL"]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
